requirements.txt
django
psycopg2
telebot
requests
django_rest_framework
